+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Changelog
 =========
+ ### Added
+ * Added a post-install console warning if terra-toolkit is included as a hard dependency
 
 5.6.2 - (September 6, 2019)
 ----------